julia_cuda/nn: compute globalId as UInt32

+++ julia_cuda/nn/nn.jl
@@ -30,7 +30,7 @@ function euclid(d_locations_ptr, d_locations_len, d_distances_ptr, d_distances_l
     d_distances = CuDeviceArray(d_distances_len, d_distances_ptr)
 
     globalId = threadIdx().x + blockDim().x *
-                (gridDim().x * (blockIdx().y - 1) + (blockIdx().x - 1))
+                (gridDim().x * (blockIdx().y - UInt32(1)) + (blockIdx().x - UInt32(1)))
     if globalId <= numRecords
         latLong = d_locations[globalId]
         d_distances[globalId] =
@@ -104,7 +104,7 @@ function main(args)
     # Execute kernel. There will be no more than (gridY - 1) extra blocks.
     @cuda ((gridX, gridY), threadsPerBlock) euclid(
         pointer(d_locations), length(d_locations), pointer(d_distances),
-        length(d_distances), numRecords, lat, lng)
+        length(d_distances), UInt32(numRecords), lat, lng)
 
     # Copy data from device memory to host memory.
     distances = Array(d_distances)
